refac: VM for RoutineDetailView
